+++ tests/test_coordinate_utils.py
@@ -274,6 +274,20 @@ class CoordinateUtilsTestCase(TestCase):
         self.assertIsNone(np.testing.assert_allclose(radius, mat['radius']))
         self.assertIsNone(np.testing.assert_allclose(theta, mat['theta']))
 
+    def test_gg_geo_gg(self):
+
+        mat = d.loadmat(MATFILE_PATH, variable_names=['test_gg_to_geo'])
+        mat = mat['test_gg_to_geo']
+
+        radius, theta = c.gg_to_geo(mat['height'], mat['beta'])
+
+        height, beta = c.geo_to_gg(radius, theta)
+
+        self.assertIsNone(
+            np.testing.assert_allclose(height, mat['height'], atol=1e-10))
+        self.assertIsNone(
+            np.testing.assert_allclose(beta, mat['beta'], atol=1e-10))
+
     def test_matrix_geo_to_base(self):
 
         theta_geo = np.linspace(1, 179, 10)